+++ xlsx_from_json/api.py
@@ -43,8 +43,8 @@ class RowFiller:
         current_row = self.start_row
 
         for row_data in rows_data:
-            current_row += row_data.get("skip_rows", 0)
-            start_column = self.start_column + row_data.get("skip_columns", 0)
+            current_row += row_data.get("rows_shift", 0)
+            start_column = self.start_column + row_data.get("columns_shift", 0)
 
             cells = self._fill_row(current_row, start_column, row_data.get("cells", []))
             row_height = max(map(attrgetter("height"), cells), default=1)